Move compiler into new TsCompiler class
